Remove calls to erlang:get_stacktrace/0
erlang:get_stacktrace/0 is deprecated and removed starting with OTP-24.
Thus, it is necessary to remove invocations to this function.

There's only one reasonable replacement if one wants to catch the call
stack of a process: querying the current stacktrace via process_info/2.
Although the return values are not equivalent, it should work
sufficiently in this context.